implementing the component  dialog of remove method

+++ src/app/main/containers/main/main.component.ts
@@ -6,6 +6,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Main } from '../../model/main';
 import { MainService } from '../../services/main.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from '../../components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-main',
@@ -20,9 +21,10 @@ export class MainComponent {
     private mainService: MainService,
     private router: Router,
     private route: ActivatedRoute,
-    private snackBar: MatSnackBar) {this.refresh();}
-
-
+    private snackBar: MatSnackBar
+  ) {
+    this.refresh();
+  }
 
   refresh() {
     this.main = this.mainService.findAll().pipe(
@@ -50,16 +52,23 @@ export class MainComponent {
   }
 
   onDelete(main: Main) {
-    this.mainService.delete(main._id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open('Curso Removido com Sucesso', 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center',
-        });
-      },
-      (error) => this.onError('Erro ao tentar remover curso!')
-    );
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.mainService.delete(main._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Curso Removido com Sucesso', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
+          },
+          (error) => this.onError('Erro ao tentar remover curso!')
+        );
+      }
+    });
   }
 }
